Refactored and fixed some bugs with opening/closing

+++ branches/summer2013/edu.rice.cs.hpc.traceviewer/src/edu/rice/cs/hpc/traceviewer/painter/BasePaintLine.java
@@ -116,7 +116,7 @@ public abstract class BasePaintLine
 				succSampleMidpoint = (int) Math.max(0, ((ptl.getTime(end)-begTime)/pixelLength)); 
 			}
 			
-			this.finishPaint(currSampleMidpoint, succSampleMidpoint, currDepth, functionName, (int) end - index + 1);
+			this.finishPaint(currSampleMidpoint, succSampleMidpoint, currDepth, functionName, end - index + 1);
 			index = end;
 		}			
 	}
